Lift max_matches limit, optionally
Summary:
This diff does three things:
* Raise the default max_matches on the backend from 50 to 5000.
* Add a new "max_matches" field in queries, thread it through to
  the search engine, and allow it to tighten the limit.
  This is boring but the bulk of the diff.
* In the web frontend, set max_matches to 100.

This gives more results while keeping things snappy.  Possibly
more than 100 would be fine too; we can play with it.

Asking the web UI to display 5000 results becomes very slow.
Because we speculatively search as you type, so the first
few characters of an intended search become a very short query
usually with many results, this is very common.  If we can
optimize this we might raise the web limit much higher too.

The biggest immediate win is in the CLI -- we can handle
5000 results in about a second.  So this opens up a wide range
of searches to do comprehensively.

(Cherry-picked from the middle of Dropbox's internal fork/branch.)