Search in repo when initiated from a file view (#6)
When a user highlights a search term in the file viewer and presses `/`
to search, we currently forget what repositories they were searching
earlier and widen back out to a search of all repositories.  Until we
can do something more sophisticated, let’s at least restrict the search
they create with `/` to the repository of the file they’re viewing.